[Feature] sc-26163 Ask user to fill credentials in non dbt case
 - Refactor user input function by inquirer
 - Refactor DataSource field to support default value and optional
 - Refactor workspace to extract DataSource and Configuration to other
   files
